[frontend] workbench creation button (#3016)
