+++ opencti-platform/opencti-front/src/private/components/data/import/ImportContent.jsx
@@ -97,18 +97,6 @@ export const importContentQuery = graphql`
         }
       }
     }
-    pendingFiles(first: 500)
-      @connection(key: "Pagination_global_pendingFiles") {
-      edges {
-        node {
-          id
-          ...WorkbenchFileLine_file
-          metaData {
-            mimetype
-          }
-        }
-      }
-    }
   }
 `;
